requirements.txt
reka-api==3.2.0
requests>=2.0.0
tqdm>=4.0.0
python-dotenv